+++ internal/sessiontest/helper_dosession_test.go
@@ -50,6 +50,7 @@ const (
 	optionClientWait
 	optionWait
 	optionPrePairingClient
+	optionLinkableKeyshareResponse
 	optionPolling
 	optionNoSchemeAssets
 )